Create Vs PC Java
It works by placing a cross in a random place and when the player puts his move it then again randomly places another cross. There is a function called the "bestmove()" which uses the "MIN MAX Algorithm" to decide the best move, but when used crashes the app. Hoping to implement it in the next update.